* feat: SEO at _document.tsx (#473)

* feat: 모든 페이지에 SEO 태그 추가

* feat: document에 공통 SEO 주입

* feat: 메타 정보 키워드 추가

* feat: apple기기에 대한 속성들 추가

* feat: add og:image

* feat: apple launch아이콘 추가 (#474)

* chore: og image 링크 수정

* feat: add apple launch logo and link tag

* chore: html lang to ko

* chore: login에서 seo 태그 제거 후 og 테스트 (#475)

* chore: SEO 컴포넌트 제거 및 기본 태그로 변경 (#476)

* chore: SEO 컴포넌트 제거

* chore: og: image 상수 변경

* chore: 애플의 다양한 기기에서의 start-up 이미지 추가 (#477)

* chore: SEO 컴포넌트 제거

* chore: og: image 상수 변경

* chore: apple icon 크기별 추가

* chore: 다양한 기기에대한 startup image link tag 추가

* refactor: Cypress Login E2E test (#478)

* feat: Cypress E2E Test Inbox Page API Mocking (#479)

* feat: Spinner에 React.memo 적용

* chore: style 주석 제거

* feat: inbox 페이지 todo api mocking

* refactor: API, ServiceWorker Sync (#481)

* feat: getTodosFromIndexedDB 함수 구현

* chore: Date type to any

* refactor: get Todo service

* chore: offline dialog 추가

* refactor: mutation

* chore: 필요없는 todo type 제거 (#482)

* Create eslint.yml

* chore: ci eslint github action 변경 (#484)

* chore: ci eslint github action 변경

* chore: remove unused cypress declaration

* refactor: update request (#485)

* refactor: update request

* refactor: update request

* chore: offline updated 컬럼 추가 (#487)

* refactor: Mutation Types (#488)

* refactor: update type 추가

* refactor: update API 타입 추가

* refactor: Update Request 타입 변경

* chore: ci 파일 주석 제거

* refactor: delete API type 작성

* chore: image Optimize (#489)

* fix: location search api end point (#491)

* refactor: Calendar (#490)

* feat: Test 페이지 기본 컴포넌트 띄우기

* feat: Calendar Navigator

* refactor: Navigator 컴포넌트 분리

* feat: robots.txt

* refactor: Calendar Header

* refactor: Calendar Body

* docs: readme에 작성한 글 링크 추가 (#493)

* refactor: offline update (#494)

* chore: og image 수정

* chore: og:image alt 추가

* refactor: onDragEnd

* chore: type

* docs: readme 시스템 아키텍처 추가

* docs: hit count (#496)

* feat: map 페이지 바운더리 추가 (#497)

* refactor: offline delete  (#498)

* refactor: delete 로직을 update로 변환

* refactor: deleteTodo props를 Optional로 변환

* feat: google calendar 연동 (#500)

* feat: google calendar 연동

* chore: url 설정

* style: google calendar 연동하기 버튼 (#501)

* chore: 지도 사진에 alt 속성 추가

* chore: import 구문 정리

* chore: add google icon

* style: google login button

* refactor: delete, update todo mutation (#502)

* fix: 휴지통 비우기시 클라이언트 상태 업데이트하지 못하는 버그 수정 (#503)

* chore: 이미지의 alt 속성 추가 (#504)

* feat: 사용자는 Map 페이지에서 현재 위치로 이동할 수 있다. (#505)

* style: 현재 위치 보여주는 기능 마크업

* feat: Map페이지에서 현재 위치로 이동하는 기능 추가

* fix: Input CLS (#506)

* chore: Recoil 제거 (#507)

* chore: useSearchSidebar hook 제거

* chore: useSidebar 훅 제거

* chore: sidebar와 관련된 atom 제거

* feat: 사용자의 마지막 사이드바 상태 기억

* chore: recoil 제거

+++ src/components/common/TodoListItem/index.tsx
@@ -22,7 +22,7 @@ interface TodoListItemProps {
   setIsFocused: Dispatch<SetStateAction<number>>;
   setIsEditing?: Dispatch<SetStateAction<number>>;
   updateTodo: UseMutateFunction<unknown, unknown, unknown, unknown>;
-  deleteTodo: UseMutateFunction<unknown, unknown, string, unknown>;
+  deleteTodo?: UseMutateFunction<unknown, unknown, string, unknown>;
 }
 
 const TodoListItem = ({
@@ -46,7 +46,12 @@ const TodoListItem = ({
   useEffect(() => {
     const deleteTodoHandler = (e: KeyboardEvent) => {
       if (e.key === 'Backspace' || e.key === 'Delete') {
-        deleteTodo(todo.id);
+        if (deleteTodo) {
+          deleteTodo(todo.id);
+          return;
+        }
+
+        updateTodo({ ...todo, deletedAt: new Date() });
       }
     };
 
@@ -55,7 +60,7 @@ const TodoListItem = ({
     }
 
     return () => document.removeEventListener('keydown', deleteTodoHandler);
-  }, [todo, deleteTodo, isFocused]);
+  }, [todo, isFocused]);
 
   useEffect(() => {
     if (isDoubleClicked && setIsEditing) setIsEditing(index);
